+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Define admin addresses for deployment (#34)
 - Mark assembly blocks as memory-safe and update deal price calculation syntax. (#33)
 - Harmonize SRLC wording. (#31)
 - Rename roles. (#29)